+++ CHANGELOG.md
@@ -17,8 +17,9 @@
 - PostgreSQL: Shorten values in hstore columns
 - PostgreSQL 11-: Avoid duplicate oid in table status (bug #1089)
 - Elasticsearch: Support dropping aliases
-- Plugins: Methods processList() and killProcess()
+- Plugins: Methods afterConnect(), processList() and killProcess()
 - New plugin: Display row numbers in select (bug #1106)
+- New plugin: Specify query timeout
 
 ## Adminer 5.3.0 (released 2025-05-04)
 - Align numeric functions right